delete file if exist

+++ fileHandling.dart
@@ -1,10 +1,15 @@
-// dart program to delete file
+// dart program to delete file if exists
 import 'dart:io';
 
 void main() {
   // open file
   File file = File('test.txt');
-  // delete file
-  file.deleteSync();
-  print('File deleted.');
+  // check if file exists
+  if (file.existsSync()) {
+    // delete file
+    file.deleteSync();
+    print('File deleted.');
+  } else {
+    print('File does not exist.');
+  }
 }
